fix: Tars no longer checks mergeable (#383)
Tars failed again, with a large number of unknown states on the tidb repository. Then I took a closer look at the logs and found that the conflicting PRs actually report errors when they are updated. The reason we couldn't update other PRs was that tars also scanned for PRs without 'status/can-merge', and those PRs returned nil because they didn't have the label, and then the loop was terminated. So this time we don't check for mergeable and update the PRs directly. Even if we encounter conflicting PRs, the update will fail and we will end up with a PR that meets the requirements or all PRs do not meet the requirements.